backend/atomicMicroservices/user/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flask-cors==5.0.1
mysql-connector-python==9.2.0
python-dotenv==1.0.1
requests==2.31.0
PyJWT==2.10.1  
SQLAlchemy==2.0.39      
